+++ src/api/resolvers/types/mutations/inputs/project/GraphQLDeleteProjectInput.ts
